Fixed annoying issue with default-terminal
I'd added a new termcap entry for tmux + italics, but obviously nothing
else has this termcap entry. So every time I SSH from a new machine I
have to download and install this termcap entry, which is a pain.

So, I just switched it to xterm, which supports the colors and italics I
want. Consequences be damned.